Added a new tool: lpi_arff
lpi_arff functions much like lpi_protoident, except that the output is in the ARFF format which is used by the WEKA machine learning software. The tool reports stats about the packet lengths and interarrival times for flows.

Very handy for doing machine learning analysis at the application layer. Thanks very much to Paweł Foremski for writing much of the code used in this tool.