Don't treat TCP keep-alives as payload-bearing
Keep-alives when the endpoint has never sent payload will use the seqno
from the SYN and contain one byte of "payload". If an endpoint sends one of
these before it sends payload, we used to treat that as the first payload
bearing packet, causing many one-way flows to be incorrectly classified.