Add -B option to all tools to control reporter thread buffering
By default, libtrace will buffer 100 results before signalling the
reporter thread to start reading them. This means that our tools
can appear to be inactive when monitoring relatively quiet live
interfaces, when in reality they are just waiting for more flows
to expire before firing up the reporter.

I've set a default of 10 completed flows, so no tool should be
inactive for too long in the presence of traffic.