Re-worked our handling of the various ffffffff game protocols
What we previously thought was Wolfenstein: Enemy Territory is now
more likely to be CS 1.6.

Wolf: ET is slightly different again, but we've managed to capture
traffic of that game and write some hopefully better rules.

Half-Life had a bit of a catch-all for any ffffffff traffic in
both directions -- this has been removed so that we can use the
unknown daily monitor to better investigate any leftovers.

Tweaked the priorities of various game rules to make sure they
are all appropriately ordered.